agelapse-desktop/requirements.txt
face_detection_tflite==0.6.0
ffmpeg_python==0.2.0
numpy==2.0.1
opencv_python==4.10.0.84
Pillow==10.4.0
pillow_heif==0.18.0
pip==24.2
PyQt5==5.15.11
PyQt5_sip==12.15.0
QtAwesome==1.3.1
